Added Permissions to Inventory, Invoice, and Loc, but commented

+++ Controllers/LocationController.cs
@@ -25,8 +25,10 @@ namespace CheckIT.API.Controllers
     public class LocationController : ControllerBase
     {
         private readonly LocRepository _repo;
+        private readonly AuthRepository _auth;
 
-        public LocationController(LocRepository repo)
+        public LocationController(LocRepository repo,
+                                AuthRepository auth)
         {
             _repo = repo;
         }
@@ -34,6 +36,15 @@ namespace CheckIT.API.Controllers
         [HttpPost("AddLocation/{locName}")]
         public async Task<IActionResult> AddLocation(string locName)
         {
+            /*
+            User user = await _auth.GetUser(int.Parse(this.User.Identity.Name));
+            
+            if (user.UserPermissions.AddLocation == false)
+            {
+                return Unauthorized();
+            }
+            */
+            
             if (ModelState.IsValid)
             {
                 var locToCreate = new Location
@@ -56,6 +67,15 @@ namespace CheckIT.API.Controllers
         [HttpGet("{Id}")]
         public async Task<IActionResult> GetLocation(int Id)//GetLocation(GetByIDDto getLocationDto)
         {
+            /*
+            User user = await _auth.GetUser(int.Parse(this.User.Identity.Name));
+            
+            if (user.UserPermissions.ViewLocation == false)
+            {
+                return Unauthorized();
+            }
+            */
+
             Location loc;
             loc = await _repo.GetLocation(Id); //GetLocation(getLocationDto.Id);
 
@@ -65,6 +85,15 @@ namespace CheckIT.API.Controllers
         [HttpGet("GetAllLocations")]
         public async Task<IActionResult> GetAllLocations()
         {
+            /*
+            User user = await _auth.GetUser(int.Parse(this.User.Identity.Name));
+            
+            if (user.UserPermissions.ViewLocation == false)
+            {
+                return Unauthorized();
+            }
+            */
+
             var locList = await _repo.GetAllLocations();
             return Ok(locList);
         }
@@ -72,6 +101,15 @@ namespace CheckIT.API.Controllers
         [HttpGet()]
         public async Task<IActionResult> GetLocations(string Name, int LocInvID)
         {
+            /*
+            User user = await _auth.GetUser(int.Parse(this.User.Identity.Name));
+            
+            if (user.UserPermissions.ViewLocation == false)
+            {
+                return Unauthorized();
+            }
+            */
+
             var LocList = await _repo.GetLocations(Name, LocInvID);
             return Ok(LocList);
         }
@@ -79,6 +117,15 @@ namespace CheckIT.API.Controllers
         [HttpDelete("DeleteLocation")] //[HttpPost("DeleteLocation/{Id}")]
         public async Task<IActionResult> DeleteLocation(int Id)
         {
+            /*
+            User user = await _auth.GetUser(int.Parse(this.User.Identity.Name));
+            
+            if (user.UserPermissions.DeleteLocation == false)
+            {
+                return Unauthorized();
+            }
+            */
+
             var deletedLocation = await _repo.DeleteLocation(Id);
             return StatusCode(201);
         }
